+++ WalletWasabi.Gui/Controls/WalletExplorer/CoinJoinTabViewModel.cs
@@ -17,6 +17,7 @@ using WalletWasabi.KeyManagement;
 using WalletWasabi.Logging;
 using WalletWasabi.Models.ChaumianCoinJoin;
 using WalletWasabi.Services;
+using WalletWasabi.Models;
 
 namespace WalletWasabi.Gui.Controls.WalletExplorer
 {
