+++ WalletWasabi.Gui/Tabs/WalletManager/LoadWalletViewModel.cs
@@ -28,6 +28,7 @@ using WalletWasabi.Hwi;
 using WalletWasabi.Hwi.Models;
 using WalletWasabi.KeyManagement;
 using WalletWasabi.Logging;
+using WalletWasabi.Models;
 
 namespace WalletWasabi.Gui.Tabs.WalletManager
 {
