+++ WalletWasabi.Gui/ViewModels/ViewModelBase.cs
@@ -6,13 +6,13 @@ using System.ComponentModel;
 using System.Linq;
 using System.Reflection;
 using WalletWasabi.Gui.ViewModels.Validation;
-using WalletWasabi.Helpers;
+using WalletWasabi.Models;
 
 namespace WalletWasabi.Gui.ViewModels
 {
 	public class ViewModelBase : ReactiveObject, INotifyDataErrorInfo
 	{
-		private List<(string, MethodInfo)> ValidationMethodCache;
+		private List<(string propertyName, MethodInfo mInfo)> ValidationMethodCache;
 		public event EventHandler<DataErrorsChangedEventArgs> ErrorsChanged;
 
 		public ViewModelBase()
@@ -20,8 +20,21 @@ namespace WalletWasabi.Gui.ViewModels
 			var vmc = Validator.PropertiesWithValidation(this).ToList();
 
 			if (vmc.Count == 0) return;
-			
+
 			ValidationMethodCache = vmc;
+			this.PropertyChanged += ValidationHandler;
+		}
+
+		private void ValidationHandler(object sender, PropertyChangedEventArgs e)
+		{
+			if (ValidationMethodCache is null) return;
+
+			foreach (var validationCache in ValidationMethodCache)
+			{
+				if (validationCache.propertyName != e.PropertyName) continue;
+
+				NotifyErrorsChanged(e.PropertyName);
+			}
 		}
 
 		public bool HasErrors => Validator.ValidateAllProperties(this, ValidationMethodCache).HasErrors;
@@ -32,7 +45,7 @@ namespace WalletWasabi.Gui.ViewModels
 
 			if (error.HasErrors)
 			{
-				return error;
+				return error.Select(p => p.Message);
 			}
 
 			return ErrorDescriptors.Empty;
