--- WalletWasabi/Helpers/ErrorDescriptor.cs
@@ -1,50 +1,0 @@
-using System;
-using System.Collections.Generic;
-
-namespace WalletWasabi.Helpers
-{
-	public class ErrorDescriptors : List<ErrorDescriptor>
-	{
-		public bool HasErrors => this.Count > 0;
-
-		public static ErrorDescriptors Empty = new ErrorDescriptors(); 
-
-		public ErrorDescriptors() : base()
-		{
-
-		}
-
-		public ErrorDescriptors(params ErrorDescriptor[] errors)
-		{
-			this.AddRange(errors);
-		}
-	}
-
-	public enum ErrorSeverity
-	{
-		Default,
-		Info,
-		Warning,
-		Error
-	}
-
-	public struct ErrorDescriptor : IEquatable<ErrorDescriptor>
-	{
-		public static ErrorDescriptor Default = new ErrorDescriptor(ErrorSeverity.Default, string.Empty);
-
-
-		public ErrorSeverity Severity { get; }
-		public string Message { get; }
-
-		public ErrorDescriptor(ErrorSeverity severity, string message)
-		{
-			this.Severity = severity;
-			this.Message = message;
-		}
-
-		public bool Equals(ErrorDescriptor other)
-		{
-			return (this.Severity == other.Severity && this.Message == other.Message);
-		}
-	}
-}
