+++ WalletWasabi/Models/ErrorSeverity.cs
@@ -1,0 +1,10 @@
+namespace WalletWasabi.Models
+{
+	public enum ErrorSeverity
+	{
+		Default,
+		Info,
+		Warning,
+		Error
+	}
+}
